Included some icons
Included some icons but have not worked with sizing them yet.

Created the Star button of the header but I don't really know how to implement the logic, I did it kind of messy for now but will work on it leater (can't seem to use variables in strings?), right now I'm just trying to make a prototype.

Will work on a better structure after the page layout is all done.