Company page view
I have put alt tags for every img element, it gets a bit messy but might be good to avoid getting any warning messages.

I have added some interfaces to the DummyModel, keep in mind these are not fully done yet, they are only used now to make it easier to try out the view but might be changed later.

I also declared some dummy constants to try out the view.

I have not yet implemented the components that should be at the top right of the page but will do that later (would like to start on the dashboard page after this and leave these components for later).

I have not yet worked with adding images to the page, I would need to see the company images first to set up the logic.

I have not made anything interactive yet since I have never worked with React, I will be prioritizing setting up the dashboard view and the model first however.

Some issues that I have is that when hovering a card the texts inside it gets a bit blurry, the same goes for when you hover the "All Stalls", I could not find a way to solve this.