Company card fixes
Fixed the css of the company card component and the ticket state component, now I'm not using any margins to make everything stick to its place but only using flex since I think it is better to do it that way.

Also fixed some sizing to make it more similar to the figma model.

Added a component AllCards that is used only to showcase, should be removed if merged to main!